+++ src/main/java/is/hi/hopur12/services/UserServiceImp.java
@@ -22,27 +22,27 @@ public class UserServiceImp implements UserService {
     
 	//Reiknar æskilegt magn af proteini per dag
 	@Override
-	public double calcProtein(double bmr) {
-		return (0.3*bmr)/4;
+	public int calcProtein(int bmr) {
+		return (int)((0.3*bmr)/4);
 	}
 	
 	//Reiknar æskilegt magn af fitu per dag
 	@Override
-	public double calcFat(double bmr) {
-		return (0.2*bmr)/9;
+	public int calcFat(int bmr) {
+		return (int)((0.2*bmr)/9);
 
 	}
 	
 	//Reiknar æskilegt magn af carbs per dag
 	@Override
-	public double calcCarbs(double bmr) {
-		return (0.5*bmr)/4;
+	public int calcCarbs(int bmr) {
+		return (int)((0.5*bmr)/4);
 	}
 	
 	// Næringargildi sem æskileg eru með tilliti til útreiknaðs
 	// BMRs, úr upplýsingum sem fengnar eru.
 	@Override
-	public Nutrition calcAll(double bmr) {
+	public Nutrition calcAll(int bmr) {
 		Nutrition nut = new Nutrition(calcProtein(bmr),calcCarbs(bmr), calcFat(bmr));
 		return nut;
 		
@@ -56,12 +56,12 @@ public class UserServiceImp implements UserService {
 	// Tekur tillit til hreyfingar og markmiða
 	@Override
 	public void calcBMR(User u) {
-		double bmrx = 0.0;
+		int bmrx = 0;
 		if(u.getGender().contentEquals("M")) {
-			bmrx = 66.5 + (13.75*u.getWeight()) + (5.003*u.getHeight()) - (6.755*u.getAge());
+			bmrx = (int)(66.5 + (13.75*u.getWeight()) + (5.003*u.getHeight()) - (6.755*u.getAge()));
 		}
 		if (u.getGender().contentEquals("F")) {
-			bmrx = 655.1 + (9.563*u.getWeight()) + (1.85*u.getHeight()) - (4.676*u.getAge());
+			bmrx = (int)(655.1 + (9.563*u.getWeight()) + (1.85*u.getHeight()) - (4.676*u.getAge()));
 		}
 		if(u.getGoal().contentEquals("1")) {
 			bmrx = bmrx-500;
@@ -74,17 +74,17 @@ public class UserServiceImp implements UserService {
 	
 	// Hendir réttu BMR í objectið user
 	// Með tilliti til hreyfingar og markmiða
-	public void checkWorkout(User u, double bmr) {
+	public void checkWorkout(User u, int bmr) {
 		if(u.getWorkout().contentEquals("1"))
-			u.setBmr(bmr*1.2);
+			u.setBmr((int)(bmr*1.2));
 		else if(u.getWorkout().contentEquals("2"))
-			u.setBmr(bmr*1.375);
+			u.setBmr((int)(bmr*1.375));
 		else if(u.getWorkout().contentEquals("3"))
-			u.setBmr(bmr*1.55);
+			u.setBmr((int)(bmr*1.55));
 		else if(u.getWorkout().contentEquals("4"))
-			u.setBmr(bmr*1.725);
+			u.setBmr((int)(bmr*1.725));
 		else if(u.getWorkout().contentEquals("5"))
-			u.setBmr(bmr*1.9);
+			u.setBmr((int)(bmr*1.9));
 	}
